requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
cachetools==5.3.3
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via pymongo
fastapi-slim==0.111.0
    # via translator-challenge (pyproject.toml)
google-api-core[grpc]==2.19.1
    # via
    #   google-cloud-core
    #   google-cloud-translate
google-auth==2.32.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-translate
google-cloud-core==2.4.1
    # via google-cloud-translate
google-cloud-translate==3.15.4
    # via translator-challenge (pyproject.toml)
googleapis-common-protos==1.63.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.64.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.64.1
    # via google-api-core
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via translator-challenge (pyproject.toml)
idna==3.7
    # via
    #   anyio
    #   requests
motor==3.5.1
    # via translator-challenge (pyproject.toml)
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-translate
protobuf==5.27.2
    # via
    #   google-api-core
    #   google-cloud-translate
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via
    #   fastapi-slim
    #   pydantic-settings
    #   translator-challenge (pyproject.toml)
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.3.4
    # via translator-challenge (pyproject.toml)
pymongo==4.8.0
    # via motor
python-dotenv==1.0.1
    # via pydantic-settings
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi-slim
typing-extensions==4.12.2
    # via
    #   fastapi-slim
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via requests
uvicorn==0.30.1
    # via translator-challenge (pyproject.toml)
uvloop==0.19.0
    # via translator-challenge (pyproject.toml)
